Moved Record interface as part of functions api
