+++ pulsar-io/aerospike/src/main/java/org/apache/pulsar/io/aerospike/AerospikeStringSink.java
@@ -19,8 +19,8 @@
 
 package org.apache.pulsar.io.aerospike;
 
+import org.apache.pulsar.functions.api.Record;
 import org.apache.pulsar.io.core.KeyValue;
-import org.apache.pulsar.io.core.Record;
 
 /**
  * Aerospike sink that treats incoming messages on the input topic as Strings
