+++ pulsar-io/twitter/src/main/java/org/apache/pulsar/io/twitter/TwitterFireHose.java
@@ -35,8 +35,8 @@ import java.io.Serializable;
 import java.util.Map;
 import java.util.Optional;
 
+import org.apache.pulsar.functions.api.Record;
 import org.apache.pulsar.io.core.PushSource;
-import org.apache.pulsar.io.core.Record;
 import org.apache.pulsar.io.core.SourceContext;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
